requirements.txt
Babel==2.9.1
blinker==1.4
click==8.0.1
cycler==0.10.0
dnspython==2.1.0
docx==0.2.4
docxtpl==0.11.5
email-validator==1.1.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fpdf==1.7.2
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
opencv-python==4.5.3.56
passlib==1.7.4
pdf2image==1.16.0
Pillow==8.3.1
PyMuPDF==1.18.15
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.8
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
redis==3.5.3
scapy==2.4.5
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.21
tk==0.1.0
Werkzeug==2.0.1
WTForms==2.3.3
